server: move the clock monotonicity logic out of server.go
Release note: None
