+++ pkg/server/BUILD.bazel
@@ -13,6 +13,7 @@ go_library(
         "authentication.go",
         "auto_tls_init.go",
         "auto_upgrade.go",
+        "clock_monotonicity.go",
         "cluster_settings.go",
         "combined_statement_stats.go",
         "config.go",
